Cleaning, adding senslab MASS conf examples, switching to Makefile.common
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/senstools/trunk@606 e59e8366-ff4a-0410-815a-bc13a6274fd4

+++ software/OS/FreeRTOS/Source/tasks.c
@@ -417,7 +417,8 @@ tskTCB * pxNewTCB;
 		#if( portSTACK_GROWTH < 0 )
 		{
 			pxTopOfStack = pxNewTCB->pxStack + ( usStackDepth - 1 );
-			pxTopOfStack = ( portSTACK_TYPE * ) ( ( ( unsigned long ) pxTopOfStack ) & ( ( unsigned long ) ~portBYTE_ALIGNMENT_MASK  ) );
+			//pxTopOfStack = ( portSTACK_TYPE * ) ( ( ( unsigned long ) pxTopOfStack ) & ( ( unsigned long ) ~portBYTE_ALIGNMENT_MASK  ) );
+			pxTopOfStack = ( portSTACK_TYPE * ) ( ( (unsigned int) pxTopOfStack ) & ( (unsigned int)  ~portBYTE_ALIGNMENT_MASK  ) );
 		}
 		#else
 		{
